User Login add

+++ tasklist/app/Http/Controllers/TasksController.php
@@ -13,11 +13,32 @@ class TasksController extends Controller
     {
         // タスクリスト一覧を取得
         $tasks = Task::all();         /* 追加 */
-
-        // タスクリスト一覧ビューでそれを表示
-        return view('tasks.index', [     /* 追加 */
-            'tasks' => $tasks,        /* 追加 */
-        ]);                                 /* 追加 */
+        
+        //dd(\Auth::check());
+        
+        //micropostcopy
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+        
+            // 認証済みユーザーを取得
+            $user = \Auth::user();
+            // ユーザーの投稿の一覧を作成日時の降順で取得
+            // （後のChapterで他ユーザーの投稿も取得するように変更しますが、現時点ではこのユーザーの投稿のみ取得します）
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            
+            // タスクリスト一覧ビューでそれを表示
+            return view('tasks.index', $data);                                 /* 追加 */
+        
+        }
+        
+        // dashboardビューでそれらを表示
+        return view('welcom', $data);
+        
+        
     }
 
     // getでtasks/createにアクセスされた場合の「新規登録画面表示処理」
@@ -44,10 +65,11 @@ class TasksController extends Controller
         $task = new Task; 
         $task->status = $request->status;
         $task->content = $request->content;
+        $task->user_id = \Auth::id();
         $task->save();
 
-        // トップページへリダイレクトさせる
-        return redirect('/');
+        // タスクリスト一覧へリダイレクトさせる
+        return redirect('/dashboard');
     }
 
     // getでtasks/（任意のid）にアクセスされた場合の「取得表示処理」
@@ -55,7 +77,13 @@ class TasksController extends Controller
     {
         // idの値でタスクリストを検索して取得
         $task = Task::findOrFail($id);
-
+        
+        // 認証済みユーザー（閲覧者）がその投稿の所有者でない場合はトップページへ
+        if (\Auth::id() != $task->user_id) {
+            // トップページへリダイレクトさせる
+            return redirect('/');
+        }
+        
         // タスクリスト詳細ビューでそれを表示
         return view('tasks.show', [
             'task' => $task,
@@ -67,6 +95,12 @@ class TasksController extends Controller
     {
         // idの値でタスクリストを検索して取得
         $task = Task::findOrFail($id);
+        
+        // 認証済みユーザー（閲覧者）がその投稿の所有者でない場合はトップページへ
+        if (\Auth::id() != $task->user_id) {
+            // トップページへリダイレクトさせる
+            return redirect('/');
+        }
 
         // タスクリスト編集ビューでそれを表示
         return view('tasks.edit', [
@@ -85,6 +119,13 @@ class TasksController extends Controller
         
         // idの値でタスクリストを検索して取得
         $task = Task::findOrFail($id);
+        
+        // 認証済みユーザー（閲覧者）がその投稿の所有者でない場合はトップページへ
+        if (\Auth::id() != $task->user_id) {
+            // トップページへリダイレクトさせる
+            return redirect('/');
+        }
+        
         // タスクリストを更新
         $task->status = $request->status;
         $task->content = $request->content;
@@ -99,8 +140,13 @@ class TasksController extends Controller
     {
         // idの値でタスクリストを検索して取得
         $task = Task::findOrFail($id);
-        // タスクリストを削除
-        $task->delete();
+        
+        // 認証済みユーザー（閲覧者）がその投稿の所有者である場合は投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            // タスクリストを削除
+            $task->delete();
+        }
+        
 
         // トップページへリダイレクトさせる
         return redirect('/');
